+++ saltar-editor/src/main/java/com/thrashplay/saltar/editor/ui/MouseSelectionController.java
@@ -25,7 +25,7 @@ public class MouseSelectionController implements UpdateableComponent {
     }
 
     @Override
-    public void update(GameObject gameObject, long delta) {
+    public void update(GameObject gameObject, float delta) {
         if (mouseTouchManager.isDown()) {
             GameObject viewport = gameObjectManager.getGameObject(GameObjectIds.ID_VIEWPORT);
             if (viewport != null) {
